Backend/requirements.txt
Django>=3
djangorestframework
gunicorn>=19.8,<19.9
whitenoise
Pillow
django-filter
django-cors-headers